Make the recipes pages open on the same page
The recipes pages were opening on a new tab and the user could return back to the home page by closing the current tab. Since a home link was included in each of the recipes, its function, returning to the home page, is no longer needed but may be reapplied in the future.